% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_uroi.R
\name{compute_greenness_time_series}
\alias{compute_greenness_time_series}
\title{Function to mask images and compute \%greenness time series (per doy)}
\usage{
compute_greenness_time_series(color = "green", x = NULL, y = NULL, folder,
  lists_files, which_images = seq_len(length(lists_files)), doy, path_base,
  name_of_analysis, main_plot, settings_mat, load_masks = FALSE,
  masks_list = NULL)
}
\arguments{
\item{color}{Color to compute the \%color time series. Default is \code{"green"}}

\item{x}{Vector of x-coordinates to read in.}

\item{y}{Vector of y-coordinates to read in.}

\item{folder}{The folder that contains the images to be read in.}

\item{lists_files}{A vector containing the names of all images to be read in.}

\item{which_images}{An optional vector to select images from \code{lists_files}.
Default is \code{seq_len(length(lists_files))}}

\item{doy}{Vector of DOYs for each image}

\item{path_base}{Base directory for saving the results}

\item{name_of_analysis}{Name of the analysis (for file name of the results)}

\item{main_plot}{Main for the plot}

\item{settings_mat}{Matrix with all settings from find_clusters}

\item{load_masks}{Boolean: Shall the masks be loaded from the disc? Otherwise, results from generate
masks can be given as masks_list}

\item{masks_list}{A list containing the masks, possibly output from \code{generate_masks()}}
}
\value{
A list containing the mean doy time series and settings
}
\description{
Function to mask images and compute \%greenness time series (per doy)
}

