% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_uroi.R
\name{create_dirs_uroi}
\alias{create_dirs_uroi}
\title{Function to create the directories for the results}
\usage{
create_dirs_uroi(path_base = NULL, name_of_analysis)
}
\arguments{
\item{path_base}{Base directory}

\item{name_of_analysis}{Name of the analysis}
}
\value{
Nothing
}
\description{
Function to create the directories for the results
}

