% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{display_eigen}
\alias{display_eigen}
\title{Display eigenimages and effect of eigenimage on the mean image}
\usage{
display_eigen(eigenimages, colormode = "Color", n_pc, images_flat_mean,
  fact = NULL, do_sum_diff = TRUE, sum1 = FALSE, save_it = FALSE,
  save_file = NULL, show_eigen = TRUE, colorspace = "rgb",
  scores = NULL)
}
\arguments{
\item{eigenimages}{eigenimages as 3D or 4D array, depending on the
color mode}

\item{colormode}{either \code{Color} or \code{Grayscale}}

\item{n_pc}{number of eigenimages}

\item{images_flat_mean}{mean image as vector}

\item{fact}{the factor, the eigenimages should be multiplied with. 
The default \code{NULL} results in weights which represent 80\% of the 
maximal and minimal score per eigenimage.}

\item{do_sum_diff}{\code{TRUE} (default): Sum and difference are displayed.
At the moment, this option works only for 
\code{sum1=FALSE}.}

\item{sum1}{\code{TRUE}: Pixel intensities are normed to sum=1, i.e., 
relative intensities are computed, default is \code{FALSE}}

\item{save_it}{\code{TRUE}: Images are saved rather than displayed, 
default is \code{FALSE}}

\item{save_file}{filename to save the images}

\item{show_eigen}{\code{TRUE} (default): Eigenimages are displayed/saved.}

\item{colorspace}{either \code{"rgb"} or \code{"hsv"}, 
default is \code{"rgb"}}

\item{scores}{scores matrix, only necessary if \code{fact==NULL} 
to compute minimal and maximal scores}
}
\description{
This function displays/saves the mean image and all eigenimages. 
	Additionally, sums and differences of mean image and
	weighted eigenimages are shown.
}

