% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{display_scores}
\alias{display_scores}
\title{Plot scores on different eigenimages}
\usage{
display_scores(scores, year, doy, complete_year = TRUE, which_year = 2000,
  which_images = NULL, n_show, save_it = FALSE, save_file = NULL,
  main = NULL, sep_plots = FALSE, plot_lowess = TRUE, ...)
}
\arguments{
\item{scores}{matrix of scores, row \code{i} contains all scores to the
\code{i}-th eigenimage}

\item{year}{vector containing the year of the image}

\item{doy}{vector containing the doy of the image}

\item{complete_year}{\code{TRUE} complete years specified in 
which_year is displayed, otherwise the scores of
\code{which_images} are displayed, default is \code{TRUE}}

\item{which_year}{vector specifying the years to be shown}

\item{which_images}{vector specifying the indices of scores to be shown}

\item{n_show}{number of shown scores time series}

\item{save_it}{\code{TRUE}: Plots are saved rather than displayed, 
default is \code{FALSE}}

\item{save_file}{filename to save the images}

\item{main}{title of the plots}

\item{sep_plots}{\code{TRUE}: Each score in a separate plot, 
default is \code{FALSE}}

\item{plot_lowess}{If \code{TRUE} (default), LOWESS estimators are plotted}

\item{...}{Further arguments}
}
\description{
This function averages the scores of all images taken at the same day
	and plots these averaged scores.
}

