% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_amos.R
\name{download_amos_func}
\alias{download_amos_func}
\title{Download image data from AMOS}
\usage{
download_amos_func(camera, year, month)
}
\arguments{
\item{camera}{Name(s) of the camera (is a number, see URL) as vector}

\item{year}{Year(s) to be downloaded as vector}

\item{month}{Month(s) to be downloaded as vector}
}
\description{
This function can be used to automatically download data from AMOS
 (\url{http://amos.cse.wustl.edu/})
}

