% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{eigen_images}
\alias{eigen_images}
\title{Function to compute  the eigenimages of a given set of images}
\usage{
eigen_images(images, n_pc = 5, varimax_rotation = FALSE, total_var = TRUE)
}
\arguments{
\item{images}{array of pixel intensities, 3rd dimension has to be the
color channels, 4th dimension the time. For grayscale images this
means that \code{dim(images) = (*,*,1,*)}, for color images 
\code{dim(images) = (*,*,3,*)}}

\item{n_pc}{number of eigenimages}

\item{varimax_rotation}{boolean: Shall varimax rotation be performed?}

\item{total_var}{Shall total variance be computed? Else 1.}
}
\description{
Images can be either grayscale images or color images. 
	The function computes the eigenimages (considering each pixel
	as different variable, i.e. no penalization) and eigenvalues.
}

