% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_uroi.R
\name{find_clusters}
\alias{find_clusters}
\title{Function to find pixel clusters in a set of images}
\usage{
find_clusters(x = NULL, y = NULL, folder, lists_files,
  which_images = seq_len(length(lists_files)), n_pc_vec, k_vec, nstart = 1,
  path_base = "", name_of_analysis = "my_analysis", save_results = FALSE,
  colormode = "Color")
}
\arguments{
\item{x}{Vector of x-coordinates to read in.}

\item{y}{Vector of y-coordinates to read in.}

\item{folder}{The folder that contains the images to be read in.}

\item{lists_files}{A vector containing the names of all images to be read in.}

\item{which_images}{An optional vector to select images from \code{lists_files}.
Default is \code{seq_len(length(lists_files))}}

\item{n_pc_vec}{Vector of number of singular vectors}

\item{k_vec}{Vector of number of clusters}

\item{nstart}{Number of new starts of the k-means}

\item{path_base}{Base directory for saving the results}

\item{name_of_analysis}{Name of the analysis (for file name of the results)}

\item{save_results}{Boolean, shall the results be saved on the disc? Default to \code{FALSE}.}

\item{colormode}{The color mode, eiteher \code{Color} or \code{Grayscale}}
}
\value{
A list containing the clusters and settings
}
\description{
Function to find pixel clusters in a set of images
}

