% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{hsv_to_rgb}
\alias{hsv_to_rgb}
\title{Transform HSV-Image to RGB-Image}
\usage{
hsv_to_rgb(images)
}
\arguments{
\item{images}{Image as 4D or 3D array}
}
\value{
4D or 3D array of same dimension as input but with rgb values
}
\description{
This function transforms a given HSV-Image (as array) into an 
	RGB-Image (as array)
}

