% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_masking.R
\name{mask_img}
\alias{mask_img}
\title{Masks images with given mask}
\usage{
mask_img(images, mask)
}
\arguments{
\item{images}{array of pixel intensities, 3rd dimension has to be the
color channels, 4th dimension the time. For grayscale images this
means that \code{dim(images) = (*,*,1,*)}, for color images 
\code{dim(images) = (*,*,3,*)}}

\item{mask}{matrix with 0/1 giving the mask}
}
\description{
Masks images with given mask, i.e., deletes pixels which are not in the mask
}

