% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{norm_img}
\alias{norm_img}
\title{Norm image to sum=1}
\usage{
norm_img(images, norm_type = "sum1")
}
\arguments{
\item{images}{Image as 4D array}

\item{norm_type}{sum1 or mean0}
}
\value{
4D array of same dimension as input but with normed values
}
\description{
This function norms an image to sum=1 or mean=0 over rows, columns and color
	channels
}

