% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{read_images}
\alias{read_images}
\title{Read images into an array}
\usage{
read_images(colormode = "Color", dim_images = NULL, folder, file_names,
  which_images, x = seq_len(dim_images[1]), y = seq_len(dim_images[2]),
  col_out = seq_len(dim_images[3]), col_in, sum1 = FALSE, norm_it = FALSE,
  norm_type = "sum1", colorspace = "rgb", relative = FALSE)
}
\arguments{
\item{colormode}{the output colormode: either "Color" (default) or "Grayscale"}

\item{dim_images}{the dimension of the output array}

\item{folder}{the path of the folder which contains the images}

\item{file_names}{a vector giving the names of the input files}

\item{which_images}{indices of the images to read, 
relative to \code{file_names}}

\item{x}{Vector of x-coordinates to read in. Default \code{NULL} results in 
a grid of all possible values.}

\item{y}{Vector of y-coordinates to read in. Default \code{NULL} results in 
a grid of all possible values.}

\item{col_out}{which color-channels shall be read in?}

\item{col_in}{which color-channels has the input image?}

\item{sum1}{\code{TRUE}: Pixel intensities are normed to sum=1, i.e. 
relative intensities are computed, default is \code{FALSE}}

\item{norm_it}{\code{TRUE}: Images are normed so that column, row and 
colour channel intensities have sum=1 or mean=0, default is \code{FALSE}}

\item{norm_type}{Type of normalization, either sum of each image = 1, 
i.e., \code{"sum1"}, or mean of each image = 0, i.e., \code{"mean0"},
default \code{"sum1"}}

\item{colorspace}{either \code{"rgb"} or \code{"hsv"}, 
default is \code{"rgb"}}

\item{relative}{\code{TRUE}: If only one colour channel shall be extracted
from a colour image, then this channel is returned relatively to the
other two channels, default is \code{FALSE}.}
}
\value{
An array containing the images. If \code{sum1=TRUE}, then the 
	third colour channel is dropped to save memory.
}
\description{
This function reads in a set of images and returns all images as an array.
}

