% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{read_images_ndvi}
\alias{read_images_ndvi}
\title{Function to read images into an array -- NDVI Version}
\usage{
read_images_ndvi(folder, file_names, which_images, x_ir, y_ir, x_red, y_red)
}
\arguments{
\item{folder}{the path of the folder which contains the images}

\item{file_names}{a vector giving the names of the input files}

\item{which_images}{indices of the images to read, 
relative to \code{file_names}}

\item{x_ir}{which x-values shall be read in for the infrared-channel?}

\item{y_ir}{which y-values shall be read in for the infrared-channel?}

\item{x_red}{which x-values shall be read in for the red-channel?}

\item{y_red}{which y-values shall be read in for the red-channel?}
}
\value{
An array containing the images.
}
\description{
Function to read images into an array -- NDVI Version
}

