% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{read_images_pseudo_ndvi}
\alias{read_images_pseudo_ndvi}
\title{Function to read images into an array -- Pseudo-NDVI Version}
\usage{
read_images_pseudo_ndvi(folder, file_names, which_images, x, y)
}
\arguments{
\item{folder}{the path of the folder which contains the images}

\item{file_names}{a vector giving the names of the input files}

\item{which_images}{indices of the images to read, 
relative to \code{file_names}}

\item{x}{which x-values shall be read in?}

\item{y}{which y-values shall be read in?}
}
\value{
An array containing the images.
}
\description{
Function to read images into an array -- Pseudo-NDVI Version
}

