% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{restore_images}
\alias{restore_images}
\title{Restore images based on eigenimages and scores}
\usage{
restore_images(eigenimages, scores, which_images, which_scores = NULL,
  colormode = "Color", images_flat_mean, display_restored = TRUE,
  relative = FALSE, col_in = 2)
}
\arguments{
\item{eigenimages}{eigen images as 3D or 4D array, depending on the
color mode}

\item{scores}{matrix with the scores on the eigenimages}

\item{which_images}{vector containing indices of images to be
restored - reduces matrix \code{scores}}

\item{which_scores}{vector containing indices of scores to use. All 
scores will be used if \code{NULL}}

\item{colormode}{either \code{Color} or \code{Grayscale}}

\item{images_flat_mean}{mean image as vector}

\item{display_restored}{\code{TRUE}: restored images are 
displayed. Default is \code{TRUE}}

\item{relative}{\code{TRUE}: Relative images. Default is \code{FALSE}}

\item{col_in}{input color channel if \code{relative==TRUE}: red=1, green=2, blue=3. Default ist 2.}
}
\value{
\code{Image} object with restored images
}
\description{
This function restores images using given eigenimages and scores
	on these eigenimages
}

