% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_classification.R
\name{revise_seasons}
\alias{revise_seasons}
\title{Function to revise classified seasons to extract unique onset days of seasons}
\usage{
revise_seasons(doys, seasons_pred)
}
\arguments{
\item{doys}{vector of given doys}

\item{seasons_pred}{vector with predicted seasons}
}
\description{
Function to revise classified seasons to extract unique onset days of seasons
}

