% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{rgb_to_hsv}
\alias{rgb_to_hsv}
\title{Transform RGB-Image to HSV-Image}
\usage{
rgb_to_hsv(images, maxColorValue = 1)
}
\arguments{
\item{images}{Image as 4D or 3D array}

\item{maxColorValue}{Maximum color value, default=1}
}
\value{
4D or 3D array of same dimension as input but with hsv values
}
\description{
This function transforms a given RGB-Image (as array) into an 
	HSV-Image (as array)
}

