% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{scale_img}
\alias{scale_img}
\title{Scale grayscale images to mean 0 and variance 1}
\usage{
scale_img(images, center = TRUE, scale = TRUE)
}
\arguments{
\item{images}{array of pixel intensities, 3rd dimension has to be the
color channels, 4th dimension the time. For grayscale images this
means that \code{dim(images) = (*,*,1,*)}, for color images 
\code{dim(images) = (*,*,3,*)}}

\item{center}{\code{TRUE} (default): Images are centered}

\item{scale}{\code{TRUE} (default): Images are standardized to variance 1}
}
\value{
The scaled images as array
}
\description{
Scale grayscale images to mean 0 and variance 1
}

