% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{scores_images}
\alias{scores_images}
\title{Function to compute the scores of the images on the eigenimages}
\usage{
scores_images(images, eigenimages, center = TRUE, norm_cols = FALSE)
}
\arguments{
\item{images}{array containing the images}

\item{eigenimages}{array containing the eigenimages}

\item{center}{\code{TRUE} (default): Images are centered around the mean.}

\item{norm_cols}{\code{TRUE}: Columns of the eigenimages are normed to length 1.
Not necessary if these are real eigenimages because then the
normalization is a result of the pca. Otherwise it could
make sense. Default is \code{FALSE}.}
}
\value{
Matrix with the scores.
}
\description{
This function projects each image on each eigenimage. Appropriate
	only if images can be given as large array. Will need another
	function if images shall be projected step by step due to 
	small working memory.
}

