% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sroi.R
\name{sroi_analysis}
\alias{sroi_analysis}
\title{Function for sROI analysis}
\usage{
sroi_analysis(folder_results = paste0(getwd(), "/"),
  folder_data = paste0(getwd(), "/"),
  name_of_analysis = substr(as.character(Sys.time()), 1, 10),
  get_timeseries = TRUE, correlate_timeseries = TRUE,
  do_evaluation = TRUE, do_strucchange = TRUE, color = "green",
  main_camera = NULL, n_pinpricks = 1, seed = 12345,
  threshold_vec = 0.9, a_vec = seq(1, 150, by = 1), b_vec = seq(1, 100, by
  = 1), lists_files, which_images = seq_len(length(lists_files)), doy,
  doy_c = doy, which_images_background = 1)
}
\arguments{
\item{folder_results}{Folder where the results will be saved}

\item{folder_data}{Folder where the data is saved}

\item{name_of_analysis}{Name of the specific analysis for a unique folder}

\item{get_timeseries}{If \code{TRUE} (default), greenness time series of 
pinpricks are extracted}

\item{correlate_timeseries}{If \code{TRUE} (default), correlation images 
are computed}

\item{do_strucchange}{If \code{TRUE} (default), optimality criteria are
computed and best mask is saved}

\item{color}{Name of color channel to be analyzed, typically \code{green} 
(default)}

\item{main_camera}{Main for the output plots}

\item{n_pinpricks}{Number of random pinpricks}

\item{seed}{Seed for the random pinpricks}

\item{threshold_vec}{Vector of thresholds for the correlation images}

\item{a_vec}{Possible spring doys}

\item{b_vec}{Possible autumn doys (counted backwards from 31.12.)}

\item{lists_files}{Vector of file names of images}

\item{which_images}{Vector of indices of images to be analyzed}

\item{doy}{Vector of DOYs for each image}

\item{doy_c}{Vector of DOYs for each image, continouing over a change of year
of more than one year is analyzed, else \code{doy}}

\item{which_images_background}{Index of background image}

\item{do_evalutaion}{If \code{TRUE} (default), correlation images and time
series are saved}
}
\description{
This function allows to carry out an entire sROI analysis for given 
webcam data
}

