% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_strucchange.R
\name{struc_dummy}
\alias{struc_dummy}
\title{Apply optimality criterion 2 on \%greenness time series}
\usage{
struc_dummy(mean_doy_norm, a_vec, b_vec, doy = NULL)
}
\arguments{
\item{mean_doy_norm}{matrix with \%greenness time series in rows}

\item{a_vec}{Vector of possible spring doys}

\item{b_vec}{Vector of possible autumn doys (backwards from 31.12.)}

\item{doy}{Optional vector of DOYs}
}
\value{
List with \code{max_corr}, the vector of OC2 values for each row of 
 \code{mean_doy_norm} and \code{wm1}, the index of the best ROI
}
\description{
Apply optimality criterion 2 on \%greenness time series
}

