% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_strucchange.R
\name{struc_ftest}
\alias{struc_ftest}
\title{Function to apply optimality criterion 1 on \%greenness time series}
\usage{
struc_ftest(mean_doy_norm, cut240 = TRUE, x_vec = NULL, uROI = TRUE,
  settings)
}
\arguments{
\item{mean_doy_norm}{matrix with \%greenness time series in rows}

\item{cut240}{Cut at DOY 240? Default is \code{TRUE}}

\item{x_vec}{Optionaly vector of DOYs}

\item{uROI}{\code{TRUE}: uROI, \code{FALSE}: sROI. Default is \code{TRUE}}

\item{settings}{MAtrix with settings, only for sROI.}
}
\value{
max_f Maximal f statistics
}
\description{
Function to apply optimality criterion 1 on \%greenness time series
}

