% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{sum1_img}
\alias{sum1_img}
\title{Transform pixel intensities to sum=1}
\usage{
sum1_img(images)
}
\arguments{
\item{images}{Image as 4D array}
}
\value{
4D array of transformed values, the third color channel is
	dropped to avoid running into collinearity problems with the
	eigenanalysis
}
\description{
This function transforms the pixel intensities of an image to sum=1
	(only over the color channel, contrary to norm_img)
}

