% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_sup_class.R
\name{sup_class}
\alias{sup_class}
\title{Function for eigenanalysis and supervised classification}
\usage{
sup_class(folder_results = paste0(getwd(), "/"),
  folder_data = paste0(getwd(), "/"),
  name_of_analysis = substr(as.character(Sys.time()), 1, 10),
  lists_files = lists_files, do_eigenanalysis = TRUE,
  do_classification = TRUE, do_classification_newyear = FALSE,
  do_masking = FALSE, load_eigen = FALSE, show_eigen = FALSE,
  varimax_rotation = FALSE, sum1 = FALSE, norm_it = TRUE,
  norm_type = "mean0", colorspace = "rgb", total_var = TRUE,
  col_in = 1:3, colormode = "Color", filename_mask = NULL,
  pfad_eigen = NULL, fact = NULL, x = NULL, y = NULL, training_year,
  test_year, season_list, n_pc = 12, year, doy, complete_year = TRUE,
  n_images_try = 10, seasons_newyear, mask = NULL)
}
\arguments{
\item{folder_results}{Folder where the results will be saved}

\item{folder_data}{Folder where the data is saved}

\item{name_of_analysis}{Name of the specific analysis for a unique folder}

\item{lists_files}{Vector of file names of images}

\item{do_eigenanalysis}{If \code{TRUE} (default), eigenanalysis is carried 
out}

\item{do_classification}{If \code{TRUE} (default), supervised classification 
is carried out}

\item{do_classification_newyear}{If \code{TRUE}, supervised 
classification of a new/test year is carried out, default is \code{FALSE}}

\item{do_masking}{If \code{TRUE}, images are masked with a given 
ROI as first step is carried out, default is \code{FALSE}}

\item{load_eigen}{If \code{TRUE}, results of eigenanalysis are 
loaded, default is \code{FALSE}}

\item{show_eigen}{If \code{TRUE}, eigenimages are save, default is 
\code{FALSE}}

\item{varimax_rotation}{If \code{TRUE}, principal components are rotated with
varimax method, default is \code{FALSE}}

\item{sum1}{If \code{TRUE}, images are normed to sum=1, default is 
\code{FALSE}}

\item{norm_it}{If \code{TRUE} (default), images are normed before the 
analysis}

\item{norm_type}{Type of norming}

\item{colorspace}{Color space (rgb, hsv, etc.)}

\item{total_var}{If \code{TRUE} (default), total variance is computed in 
eigenanalysis}

\item{col_in}{Color channels to be read in, default \code{1:3} corresponds
to red, blue and green}

\item{colormode}{The color mode}

\item{filename_mask}{File name of mask / ROI}

\item{pfad_eigen}{Path of results of eigenanalysis to be loaded}

\item{fact}{Weighting factor for sum and diff of mean and eigenimages, 
default is \code{NULL} and then quantiles of the scores are used as weights}

\item{x}{vector of x-coordinates of analyzed pixels}

\item{y}{vector of y-coordinates of analyzed pixels}

\item{training_year}{The training year}

\item{test_year}{The test year}

\item{season_list}{List specifying the true seasons}

\item{n_pc}{Number of principal components / eigenimages}

\item{year}{Vector of year for each image}

\item{doy}{Vector of DOYs for each image}

\item{complete_year}{If \code{TRUE} (default), the complete year is analysed,
else the first \code{n_images_try} are analyzed (only for testing)}

\item{n_images_try}{See complete_year}

\item{seasons_newyear}{Vector with true seasons of the test year}

\item{mask}{Do not edit, mask is loaded, if needed}
}
\value{
An entire supervised classification analysis is carried out.
 All results are saved on the disc, the necessary folders are created.
}
\description{
This function allows to carry out an eigenanalysis of given webcam data
 and a supervised classification of DOYs with respect to seasons. 
 Example code can be downloaded at 
 \url{http://bothmann.userweb.mwn.de/dissertation.html}.
}

