% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_amos.R
\name{amos_uroi_wrap}
\alias{amos_uroi_wrap}
\title{Wrapper function for analyzing data from AMOS (http://amos.cse.wustl.edu/)}
\usage{
amos_uroi_wrap(camera, year_analysis, months_analysis, hour_analysis,
  do_mean_greenness = TRUE, do_strucchange = TRUE, testmode = FALSE,
  only_download = FALSE, folder_results = getwd(),
  name_of_analysis = substr(as.character(Sys.time()), 1, 10),
  folder_data = getwd(), n_pc_vec = 12, k_vec = 4:10, nstart = 2,
  save_results = TRUE, save_masks = TRUE, masks_type = ".jpg",
  a_vec = seq(1, 150, by = 1), b_vec = seq(1, 100, by = 1), ...)
}
\arguments{
\item{camera}{Name of the camera (is a number, see URL)}

\item{year_analysis}{Year(s) to be analyzed as vector}

\item{months_analysis}{Month(s) to be analyzed as vector}

\item{hour_analysis}{Hour(s) to be analyzed as vector}

\item{do_mean_greenness}{If \code{TRUE} (default), time series of percentage 
greenness inside the ROIs are computed}

\item{do_strucchange}{If \code{TRUE} (default), points of structural change 
are searched for and OC values computed}

\item{testmode}{If \code{TRUE}, only 10 images for testing are analyzed, 
default is \code{FALSE}}

\item{only_download}{If \code{TRUE}, images are downloaded but not analyzed,
default is \code{FALSE}}

\item{folder_results}{Base folder where the results will be saved, default is wd}

\item{name_of_analysis}{Name of subfolder for the results of this run,
default is the date as \code{yyyy-mm-dd}}

\item{folder_data}{Folder where the data will be saved, default is wd}

\item{n_pc_vec}{Vector of numbers of eigenimages}

\item{k_vec}{Vector of numbers of clusters for k-means}

\item{save_results}{\code{TRUE} if results of clustering shall be saved}

\item{save_masks}{\code{TRUE} if masks shall be saved}

\item{masks_type}{File name extension of masks, default is \code{.jpg}}

\item{a_vec}{Possible spring doys}

\item{b_vec}{Possible autumn doys (counted backwards from 31.12.)}

\item{n_start}{Number of iterations k-means}
}
\description{
Wrapper function for analyzing data from AMOS (http://amos.cse.wustl.edu/)
}

