% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_uroi_part.R
\name{create_dirs_uroi}
\alias{create_dirs_uroi}
\title{Create directories for the results}
\usage{
create_dirs_uroi(path_base = NULL, name_of_analysis)
}
\arguments{
\item{path_base}{Base directory}

\item{name_of_analysis}{Name of the analysis}
}
\value{
Directories \code{path_base/name_of_analysis/mask_timeseries} and
 \code{path_base/name_of_analysis/masks} are created
}
\description{
This function creates the specified directories
}

