% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{display_eigen}
\alias{display_eigen}
\title{Display the eigenimages and effect of a multiple on the mean image}
\usage{
display_eigen(eigenimages, colormode = "Color", n_pc, images_flat_mean,
  fact = NULL, do_sum_diff = TRUE, sum1 = FALSE, save_it = FALSE,
  save_file = NULL, show_eigen = TRUE, colorspace = "rgb",
  scores = NULL)
}
\arguments{
\item{eigenimages}{eigen images as 3D or 4D array, depending on the
color mode}

\item{colormode}{either \code{Color} or \code{Grayscale}}

\item{n_pc}{number of eigenimages}

\item{images_flat_mean}{mean image as vector}

\item{fact}{the factor, the eigenimages should be multiplied with}

\item{do_sum_diff}{\code{TRUE}: Sum and difference are displayed - 
default.# At the moment, this option works only for 
\code{sum1=FALSE}}

\item{sum1}{\code{TRUE}: Pixel intensities are normed to sum=1, i.e. 
relative intensities are computed, default is \code{FALSE}}

\item{save_it}{\code{TRUE}: Images are saved rather than displayed, 
default is \code{FALSE}}

\item{save_file}{filename to save the images}

\item{show_eigen}{\code{TRUE}: Eigenimages are displayed/saved,
default is \code{TRUE}}

\item{colorspace}{either \code{"rgb"} or \code{"hsv"}, 
default is \code{"rgb"}}

\item{scores}{scores matrix, only necessary if \code{fast==NULL} 
to compute minimal and maximal scores}
}
\description{
This function displays the mean image and all eigenimages. 
	Additionaly, the sum and difference of the mean image and a 
	suitable multiple of the images is shown optionally.
}

