% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_strucchange.R
\name{dummy_time}
\alias{dummy_time}
\title{Help function for computation of dummy time series}
\usage{
dummy_time(x, a, b, doy = NULL)
}
\arguments{
\item{x}{Vector of time points}

\item{a}{First break point at a}

\item{b}{Second break point at 365-b}

\item{doy}{Vector of doys}
}
\description{
Help function for computation of dummy time series
}

