% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_uroi.R
\name{overlay_masks_onref}
\alias{overlay_masks_onref}
\title{Overlay list of mask images on reference image with transparency alpha}
\usage{
overlay_masks_onref(masks_list, ref_image, alpha = 0.1)
}
\arguments{
\item{masks_list}{List of masks, preferably output from \code{\link{generate_masks}}}

\item{ref_image}{Reference image for the overlay}

\item{alpha}{Transparency factor}
}
\value{
An \code{Image} object containing all overlays.
}
\description{
Overlay list of mask images on reference image with transparency alpha
}
\author{
Michael Matiu
}

