% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_classification.R
\name{revise_seasons}
\alias{revise_seasons}
\title{Revise classified seasons to extract unique onset days of seasons}
\usage{
revise_seasons(doys, seasons_pred)
}
\arguments{
\item{doys}{vector of given doys}

\item{seasons_pred}{vector with predicted seasons}
}
\value{
List with revised time series of seasons and extracted unique season
 onset dates.
}
\description{
Revise classified seasons to extract unique onset days of seasons
}

