% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_input.R
\name{scale_img}
\alias{scale_img}
\title{Function to scale a Grayscale image to mean 0 and variance 1}
\usage{
scale_img(images, center = TRUE, scale = TRUE)
}
\arguments{
\item{images}{array of pixel intensities, 3rd dimension has to be the
color channels, 4th dimension the time. For grayscale images this
means that \code{dim(images) = (*,*,1,*)}, for color images 
\code{dim(images) = (*,*,3,*)}}

\item{center}{boolean, shall the images be centered? 
default is \code{TRUE}}

\item{scale}{boolean, shall the images be standardized to variance 1? 
default is \code{TRUE}}
}
\description{
Function to scale a Grayscale image to mean 0 and variance 1
}

