% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_eigenanalysis.R
\name{scores_images}
\alias{scores_images}
\title{Function to compute the scores of the images on the eigenimages}
\usage{
scores_images(images, eigenimages, center = TRUE, norm_cols = FALSE)
}
\arguments{
\item{images}{array containing the images}

\item{eigenimages}{array containing the eigenimages}

\item{center}{Shall the images be centered around the mean? 
Default is \code{TRUE}}

\item{norm_cols}{Shall columns of the eigenimages be normed to length 1?
Not necessary if these are real eigenimages because then the
normalization is a result of the pca. Otherwise it could
make sense.}
}
\description{
This function projects each image on each eigenimage. Appropriate
	only if images can be given as large array. Will need another
	function if images shall be projected step by step due to 
	small working memory. Currently, images are centered around the
	mean image based on the images. Is it better to center around the
	images used to extract the eigenimages?
}

