% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_strucchange.R
\name{struc_dummy}
\alias{struc_dummy}
\title{Function to apply optimality criterion 2 on \%greenness time series}
\usage{
struc_dummy(mean_doy_norm, a_vec, b_vec, doy = NULL)
}
\arguments{
\item{mean_doy_norm}{matrix with \%greenness time series in rows}

\item{a_vec}{Vector of possible spring doys}

\item{b_vec}{Vector of possible autumn doys (backwards from 31.12.)}

\item{doy}{Optionaly vector of DOYs}
}
\value{
max_corr Maximal correlations
}
\description{
Function to apply optimality criterion 2 on \%greenness time series
}

